Moved CV to a new location + typo fix
